requirements.txt
fastapi==0.112.0
uvicorn==0.30.1
aiogram==3.21.0
python-dotenv==1.0.1
requests==2.28.2
